+++ analyzer/src/main/java/up/visulog/analyzer/CountCommitsPerAuthorPlugin.java
@@ -66,17 +66,17 @@ public class CountCommitsPerAuthorPlugin extends Plugin {
 
         public Webgen.Graph[] getResultAsGraphArray() {
             ArrayList<String> labels = new ArrayList<String>();
-            ArrayList<Integer> data = new ArrayList<Integer>();
+            ArrayList<Double> data = new ArrayList<Double>();
             for (var item : commitsPerAuthor.entrySet()) {
                 labels.add(item.getKey());
-                data.add(Integer.valueOf(item.getValue()));
+                data.add(Double.valueOf(item.getValue()));
             }
             String[] labelsArray = new String[labels.size()];
-            int[] dataArray = new int[data.size()];
+            double[] dataArray = new double[data.size()];
             for(int i=0; i<labelsArray.length; i++) labelsArray[i] = labels.get(i);
             for(int i=0; i<dataArray.length; i++) dataArray[i] = data.get(i);
             return new Webgen.Graph[]{
-                new Webgen.BarGraph("<i class=\"fas fa-pen-fancy\"></i>","Commits per author - Bar", labelsArray, new String[]{"Commits per author - Bar"}, new int[][]{dataArray}, new Color[][]{Webgen.generateRandomColorArray(dataArray.length)})
+                new Webgen.BarGraph("<i class=\"fas fa-pen-fancy\"></i>","Commits per author - Bar", labelsArray, new String[]{"Commits per author - Bar"}, new double[][]{dataArray}, new Color[][]{Webgen.generateRandomColorArray(dataArray.length)})
             };
         }
